Clean up GroupsController
