+++ SleekChat.Data/Helpers/ValidationHelper.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Text.Json;
 
 namespace SleekChat.Data.Helpers
 {
